chore: rename debug tab content files
