+++ ui/pages/settings-v2/debug-tab/debug-tab-content.stories.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import DebugTabContent from './index';
+
+const DebugTabStory = {
+  title: 'Pages/Settings/DebugTab',
+
+  component: DebugTabContent,
+};
+
+export const DefaultStory = ({ variant, address }) => <DebugTabContent />;
+
+DefaultStory.storyName = 'Default';
+
+export default DebugTabStory;
