+++ ui/pages/settings-v2/debug-tab/debug-tab-toggle-row.test.tsx
@@ -3,7 +3,7 @@ import configureMockStore from 'redux-mock-store';
 import thunk from 'redux-thunk';
 import { renderWithProvider } from '../../../../test/lib/render-helpers-navigate';
 import mockState from '../../../../test/data/mock-state.json'; // Adjust the path based on your actual file location
-import ToggleRow from './developer-options-toggle-row-component';
+import ToggleRow from './debug-tab-toggle-row';
 
 describe('ToggleFeatureRow Component', () => {
   const mockTitle = 'Test Title';
