+++ ui/pages/settings-v2/debug-tab/debug-tab-toggle-row.tsx
