--- ui/pages/settings-v2/debug-tab/developer-options-tab.stories.tsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import DeveloperOptionsTab from './index';
-
-const DeveloperOptionsTabStory = {
-  title: 'Pages/Settings/DeveloperOptionsTab',
-
-  component: DeveloperOptionsTab,
-};
-
-export const DefaultStory = ({ variant, address }) => <DeveloperOptionsTab />;
-
-DefaultStory.storyName = 'Default';
-
-export default DeveloperOptionsTabStory;
